Removing unreliable method
This doesn't work because the overrides on the class are not torn down after
each example is run. This was created so that we could mock the current_user
instance variable to return the PRs that we specify. Finding a better that to
mock the PRs of the current user will remove the need to mock the current_user.