WIP for fetching just cursors for issues from the Github API
This was done when I thought the github API returned more than 1000 results and
we would need to break the import task into many background jobs. This fetches a
list of cursors from the search results, which we can then send to unique jobs
to fetch the total issues. This will be useful if our import process proves to
be unreliable and needs to be decomposed into separate jobs